requirements.txt
Django==1.7
django-braces==1.8.0
django-filter==0.9.2
django-oauth-toolkit==0.8.1
djangorestframework==3.1.1
Markdown==2.6.2
oauthlib==0.7.2
PyJWT==1.1.0
python-openid==2.2.5
python-social-auth==0.2.9
requests==2.7.0
requests-oauthlib==0.5.0
simplejson==3.6.5
six==1.9.0